requirements.txt
flwr>=1.0.0
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
tensorflow>=2.8.0
tensorflow-privacy>=0.7.0
torch>=1.10.0
pyyaml>=6.0
matplotlib>=3.5.0
seaborn>=0.11.0
cryptography>=37.0.0
protobuf>=3.19.0
psutil>=5.9.0
phe>=1.5.0

